added read validation + major refactor
